Remove Maven warnings
This commit removes the warnings that Maven emits when building the
effective model of the project. This boils down to locking down plugin
versions.

Because there is no version for javaformat, Maven could decide to use
the latest, which requires extra care to build with Java 8. This commit
upgrades explicitly to the latest and configure it to be compatible
with Java 8