Do not apply Javaformat by default
* Do not apply javaformat by default

This commit removes the invocation of spring-javaformat:apply by
default. Doing so makes the validate goal irrelevant as apply is
going to reformat the code if necessary.

The idea of the plugin is that it should prevent you from pushing
something that has formatting violations. Running the build, run the
apply goal manually, and amending the commit is the idiomatic way of
using this plugin.

* Apply formatting